+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+* Allow `tasks.py` to be imported when Celery is not installed. This allows
+  auto-discovery by other task systems such as Huey to succeed.
+
 Version 3.4.0 (2020-04-13)
 --------------------------
 * Signals that emails have been put into the queue.